moved little filter stuff to littleFilters.ts
